+++ port/platform/esp-idf/src/u_port.c
@@ -133,7 +133,7 @@ void uPortDeinit()
 }
 
 // Get the current tick converted to a time in milliseconds.
-int64_t uPortGetTickTimeMs()
+int32_t uPortGetTickTimeMs()
 {
     return esp_timer_get_time() / 1000;
 }
